archive: do not double-close archives on ctf_arc_write() close() error
An error from close() at archive write time is actually fairly likely as
errors from close() go. After such errors, the state of the fd is
unclear: it may still be open but it certainly isn't safe to try and
close it again.  Restructure things to not close in this situation
(with a bit more repetition to avoid unlinking an open file if close()
did succeed in closing the file, since some OSes dislike that).

Spotted by Indu.