dump: introduce ctf_str_append_noerr and use it in dump
ctf-dump makes heavy use of ctf_str_append, ignoring error returns.
This is likely to coredump on error.

Rather than massively uglifying the code handling OOM in all these
cases, introduce a variant of ctf_str_append that simply returns the old
string on OOM, and use it: this is only a dump, and missing a bit of
text on OOM is likely to be a much better outcome than coredumping.
(Fix up the single unadorned strdup at the same time.)