ctf_add_type: return existing type using a variable that's actually defined
The code to return existing structs/unions/enums was broken because we
were referencing an uninitialized variable to get the vlen rather than
the variable we'd just initialized in the previous line.

Move the declaration of the variable we mistakenly used into an inner
scope so we can't mistakenly use it again.