create: forwards are always in the namespace of their referent
The C namespace a forward is located in is always the same as the
namespace of the corresponding complete type: 'struct foo' is in the
struct namespace and does not collide with, say, 'union foo'.

libctf allowed for this in many places, but inconsistent: in particular,
forward *addition* never allowed for this, and was interning forwards in
the default namespace, which is always wrong, since you can only forward
structs, unions and enums, all of which are in their own namespaces in
C.

Forward removal needs corresponding adjustment to remove the names form
the right namespace, as does ctf_rollback.