Fix lifetimes of CTF sections.
When I added ctf_arc_bufopen I forgot the cardinal rule of CTF section
lifetimes (and, for that matter, C data structures in general): the
allocator must always free.

So, rather than unconditionally freeing the symsect and freeing the
strsect when ctf_open_bfd calls for it (which literally nothing is ready
for, and which leads to double-frees in all users), never free the
strsect and free the symsect when ctf_open_bfd calls for it.